pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.bandit]
# B101 tests the use of assert
# B301 and B403 test the use of pickle
skips = ["B101", "B301", "B403"]
exclude_dirs = ["tests", ".tox", ".venv"]

pythonpath = "src/"

[project]
name = "trace-gfn"
readme = "README.md"
version = "0.1.0"
requires-python = ">=3.10,<3.12"
description = "gflownet"

dependencies = [
    "antlr4-python3-runtime==4.9.3",
    "botorch==0.11.3",
    "certifi==2024.8.30",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "cvxopt==1.3.2",
    "filelock==3.16.0",
    "fsspec==2024.9.0",
    "gitdb==4.0.11",
    "GitPython==3.1.43",
    "gpytorch==1.12",
    "idna==3.10",
    "jaxtyping==0.2.34",
    "Jinja2==3.1.4",
    "joblib==1.4.2",
    "linear-operator",
    "MarkupSafe",
    "mpmath",
    "multipledispatch",
    "networkx",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "opt-einsum==3.3.0",
    "pandas",
    "pillow==10.2.0",
    "platformdirs==4.3.3",
    "protobuf==5.28.1",
    "psutil==6.0.0",
    "pyarrow==17.0.0",
    "pyparsing==3.1.4",
    "pyro-api==0.1.2",
    "pyro-ppl==1.9.1",
    "pytorch-lightning",
    "PyYAML==6.0.2",
    "rdkit==2024.3.5",
    "requests==2.32.3",
    "scikit-learn==1.2.2",
    "scipy==1.14.1",
    "sentry-sdk==2.14.0",
    "setproctitle==1.3.3",
    "six==1.16.0",
    "smmap==5.0.1",
    "sympy==1.13.2",
    "threadpoolctl==3.5.0",
    "torch==2.1.2",
    "torch_geometric==2.4.0",
    "torchaudio",
    "torchtext",
    "torchvision",
    "tqdm",
    "triton",
    "typeguard",
    "typing_extensions",
    "urllib3",
    "wandb"
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux'" },
]
torchaudio = [
  { index = "pytorch-cu121", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
